README.md

+++ fingerprinting-async/run.sh
@@ -1,12 +1,27 @@
 #!/bin/bash
+
+if [ "$#" -ne 3 ]; then
+    echo "Usage: $0 <IP_ADDRESS> <PORT> <INTERFACE>"
+    exit 1
+fi
+
+IP_ADDRESS=$1
+PORT=$2
+INTERFACE=$3
+
 mkdir -p results_local
-for i in {82..100} ;
-do
+
+for i in {1..100}; do
     dd if=/dev/urandom of=random_memory bs=4096 count=32
     dd if=/dev/urandom of=random_memory3 bs=4096 count=32
+    
     echo -n "$i;" >> results_local.txt
-    python3 attacker.py -f random_memory -f libc-2.23.so.32 -f random_memory3 -f libc-2.19.so.32  -h 10.27.142.89 -p 7777 -d enp7s0 -n 20 >> results_local.txt
+    
+    python3 attacker.py -f random_memory -f libc-2.23.so.32 -f random_memory3 -f libc-2.19.so.32 -h $IP_ADDRESS -p $PORT -d $INTERFACE -n 20 >> results_local.txt
+    
     python3 evaluate.py log.csv >> results_local.txt
-    echo "done with $i";
+    
+    echo "done with $i"
+    
     cp log.csv results_local/$i.csv
 done
